Migrate to Service Provider Fixes #1
