+++ README.md
@@ -1,0 +1,30 @@
+## Plugin "Matomo
+This Plugin has two fields integrated.
+
+* Copy in the first field your tracking-code
+```
+<script>
+    var _paq = window._paq = window._paq || [];
+    ...
+</script>
+<noscript>
+    ...
+</noscript>
+```
+* Copy in the second field your optout-code
+```
+<iframe src="your_link" style="border: 0; height: 200px; width: 600px;">
+</iframe>
+```
+
+* Set in your article the short code {matomo_opt_out}
+* The plugin replace the short code with your optout-code
+
+### Your Matomo Codes
+You must generated the tracking-code and optout-code in your matomo-system, not in this plugin.
+
+### GDPR / DSGVO
+Please observe the respective national laws. This plugin does not check whether cookies are used or whether consent has been given.
+
+## Download an more information
+[it-conserv.de](https://it-conserv.de/joomla/module/openhours.html)
