--- index.html
@@ -1,1 +1,0 @@
-<html><body></body></html>
