+++ services/provider.php
@@ -1,0 +1,52 @@
+<?php
+
+/**
+* Joomla.Plugin - itcs Matomo
+* ------------------------------------------------------------------------
+* @package     itcs Matomo
+* @author      it-conserv.de
+* @copyright   2023 it-conserv.de
+* @license     GNU/GPLv3 <http://www.gnu.org/licenses/gpl-3.0.de.html>
+* @link        https://it-conserv.de
+* ------------------------------------------------------------------------
+*/
+
+\defined('_JEXEC') or die;
+
+use Joomla\CMS\Extension\PluginInterface;
+use Joomla\CMS\Factory;
+use Joomla\CMS\Plugin\PluginHelper;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+use Joomla\Event\DispatcherInterface;
+use ITCS\Plugin\System\ItcsMatomo\Extension\ItcsMatomo;
+
+return new class () implements ServiceProviderInterface {
+    /**
+     * Registers the service provider with a DI container.
+     *
+     * @param   Container  $container  The DI container.
+     *
+     * @return  void
+     *
+     * @since   4.3.0
+     */
+    
+    public function register(Container $container)
+    {
+        $container->set(
+            PluginInterface::class,
+            function (Container $container) {
+                $dispatcher = $container->get(DispatcherInterface::class);
+
+                $plugin = new ItcsMatomo(
+                    $dispatcher,
+                    (array) PluginHelper::getPlugin('system', 'itcs_matomo')
+                );
+                $plugin->setApplication(Factory::getApplication());
+
+                return $plugin;
+            }
+        );
+    }
+};
